+++ Source/Examples/PropertyGrid/CustomFactoryDemo/Factories/CustomPropertyGridControlFactory.cs
@@ -23,8 +23,7 @@ namespace CustomFactoryDemo
     /// </summary>
     public class CustomPropertyGridControlFactory : PropertyGridControlFactory
     {
-        public override FrameworkElement CreateControl(PropertyItem pi, PropertyControlFactoryOptions options, 
-            object instance)
+        public override FrameworkElement CreateControl(PropertyItem pi, PropertyControlFactoryOptions options, object instance = null)
         {
             // Check if the property is of type Range
             if (pi.Is(typeof(Range)))